BitrixCorp/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.2
faker==28.4.1
